Next

+++ pkg/spammer/spammer.go
@@ -411,7 +411,7 @@ func (s *Spammer) doSpam(ctx context.Context, currentProcessID uint32) error {
 				s.LogDebugf("On Create Total Alias Available: %d", len(s.accountSender.AliasOutputs()))
 				s.LogDebugf("On Create Loop: %t", s.valueLoopTransitionAlias)
 				
-				if !(s.valueLoopTransitionAlias && uint64(len(s.accountSender.AliasOutputs()) < s.valueLoopTotalAliasNumber)) {
+				if !(s.valueLoopTransitionAlias && len(s.accountSender.AliasOutputs()) < int(s.valueLoopTotalAliasNumber)) {
 					s.LogDebugf("Number of Alias Outputs before creating: %d", len(s.accountSender.AliasOutputs()))
 					if err := s.aliasOutputCreate(ctx, s.accountSender, s.valueAliasPayloadSize, outputStateNamesMap[s.outputState]); err != nil {
 						logDebugStateErrorFunc(s.outputState, err)
